Allow building multiple versions at once
These all would use the same source, so the use case is when you want
the same built documentation to appear in multiple places.

Note that we do a separate build for each version. This is
inefficient (assuming that the desired end result is for the different
versions to be identical), but was an easy-to-implement solution to add
convenience for building multiple versions of documentation with short
build times (i.e., rather than requiring you to rerun build_docs
multiple times). If this multiple-versions-at-once option starts to be
used a lot, we could reimplement it to build just one version then copy
the builds to the other versions (if that gives the correct end result).