With Docker: Make a sym link mapping the docker mountpoint to home
The need for this is subtle: For CTSM, the documentation build invokes
'git lfs pull'. However, when doing the documentation build from a git
worktree, the .git directory is replaced with a text file giving the
absolute path to the parent git repository, e.g., 'gitdir:
/Users/sacks/ctsm/ctsm0/.git/worktrees/ctsm5'. So when trying to execute
a git command from within the Docker image, you get a message like,
'fatal: not a git repository:
/Users/sacks/ctsm/ctsm0/.git/worktrees/ctsm5', because in Docker-land,
this path doesn't exist. To work around this problem, we create a sym
link in Docker's file system with the appropriate mapping. For example,
if the local file system's mount-point is /path/to/foo, then we create a
sym link at /path/to/foo in Docker's file system, pointing to the home
directory in the Docker file system.