Attempt to run 'make fetch-images' before building documentation
CTSM requires this to be run. We used to do that as part of the build
process, by having Makefile rules ensuring that the 'fetch-images'
target is executed before html or latexpdf. However, that can cause
problems if building via Docker, e.g., if you set up your repository via
ssh keys (since the Docker image won't be able to use those). So now I
am changing the process so that 'make fetch-images' is run by build_docs
before launching the Docker image.