Only fetch images if requested
I realized a problem with the previous implementation: if the image
fetch *is* supposed to happen for a repository, but there is an error in
trying to fetch the images, then the process just silently continues. So
I think it's best to require this to be explicitly requested.

In CTSM I'm handling this by having a little build_docs wrapper shell
script that always includes the '--fetch-images' option.