With Docker: Actually, never create the sym link
I realized that, now that we do the fetch-images step prior to docker
run, I don't think we'll ever need the sym link. Removing that
simplifies the code and removes a possible source of issues (e.g., if
the path where we're trying to create the symlink already exists, or
there is some other issue with creating the symlink).