Remove the --fetch-images option
None of the things I tried work robustly anymore to fetch images for the
CTSM repository. It seems like a git lfs pull or git lfs fetch doesn't
work from within a python-spawned subprocess for some reason that I
can't determine. So I'm going to remove the attempt to do this and just
leave it up to the user to do this git lfs fetch outside of this tool.